Split lib/gsuite (#44345)
* Inline exports

* Split out formatPrice

* Split out getAnnualPrice

* Split out getMonthlyPrice

* Split out product slug exports

* Split out getLoginUrlWithTOSRedirect

* Split out canUserPurchaseGSuite

* Split out canDomainAddGSuite

* Split out getGSuiteMailboxCount

* Split out hasGSuiteWithUs

* Split out hasGSuiteWithAnotherProvider

* Split out hasPendingGSuiteUsers

* Split out supported domain functions

* Split out getEligibleGSuiteDomain

* Sort exports

* Add package.json file
