+++ client/lib/gsuite/get-eligible-gsuite-domain.js
@@ -1,0 +1,35 @@
+/**
+ * External dependencies
+ */
+import { get, sortBy } from 'lodash';
+
+/**
+ * Internal dependencies
+ */
+import { canDomainAddGSuite } from './can-domain-add-gsuite';
+import { getGSuiteSupportedDomains } from './gsuite-supported-domain';
+
+/**
+ * Retrieves the first domain that is eligible to G Suite in this order:
+ *
+ *   - The domain from the site currently selected, if eligible
+ *   - The primary domain of the site, if eligible
+ *   - The first non-primary domain eligible found
+ *
+ * @param {string} selectedDomainName - domain name for the site currently selected by the user
+ * @param {Array} domains - list of domain objects
+ * @returns {string} - the name of the first eligible domain found
+ */
+export function getEligibleGSuiteDomain( selectedDomainName, domains ) {
+	if ( selectedDomainName && canDomainAddGSuite( selectedDomainName ) ) {
+		return selectedDomainName;
+	}
+
+	// Orders domains with the primary domain in first position, if any
+	const supportedDomains = sortBy(
+		getGSuiteSupportedDomains( domains ),
+		( domain ) => ! domain.isPrimary
+	);
+
+	return get( supportedDomains, '[0].name', '' );
+}
