+++ client/lib/gsuite/get-gsuite-mailbox-count.js
@@ -1,0 +1,8 @@
+/**
+ * External dependencies
+ */
+import { get } from 'lodash';
+
+export function getGSuiteMailboxCount( domain ) {
+	return get( domain, 'googleAppsSubscription.totalUserCount', 0 );
+}
