+++ client/lib/gsuite/package.json
@@ -1,0 +1,3 @@
+{
+	"sideEffects": false
+}
